# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core backend service for BookEat.\n\nEvery response is wrapped in a JSON envelope — {\"data\": ...} on\nsuccess, {\"error\": \"...\"} on failure — except /.well-known/jwks.json,\nwhich returns a raw JWKS document.",
        "title": "BookEat backend-core API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates by email + password and returns a token pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.tokenPairResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Revokes the given refresh token so it can no longer be used.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log out",
                "parameters": [
                    {
                        "description": "Refresh token to revoke",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\\\"data\\\":{\\\"ok\\\":true}}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/request": {
            "post": {
                "description": "Generates a one-time code and delivers it to the phone. Rate-limited\n(per-minute and per-hour); over the limit returns 422. The response\n\"code\" field is populated only when AUTH_OTP_DEV_EXPOSE=true.\n\nEvery 422 carries a machine-readable \"code\" in the error envelope —\nbranch on it, never on the message: otp_invalid_phone (the number is\nunusable, point at the field), otp_rate_limited_minute and\notp_rate_limited_hour (both carry a Retry-After header in seconds;\nit is the full window, an upper bound, not the exact time left).\n\nWhen no delivery channel accepts the code the answer is a plain 500\nwith the generic body — identical for a number nobody can reach and\nfor a provider outage. That is deliberate: a response that varied by\nchannel would let anyone enumerate which numbers have Telegram or\nWhatsApp. The per-channel reasons live in our logs, phone masked.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Request an OTP code",
                "parameters": [
                    {
                        "description": "Phone number",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.otpRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.otpRequestedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "validation failed / rate limited",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "the code could not be delivered on any channel",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/verify": {
            "post": {
                "description": "Verifies the latest active code for the phone. On success, finds or\ncreates the user and returns a token pair. Wrong/expired codes and\ntoo many attempts return 401.\n\nThe 401 carries one of two codes. \"otp_invalid\" means the code was\nnot accepted and covers three cases on purpose — wrong, expired, and\nno active code at all: telling them apart would let anyone with a\nphone number detect when its owner is mid-login. \"otp_too_many_attempts\"\nmeans the code is dead from wrong guesses and only a new code helps.\nThe 422 codes are otp_invalid_phone and otp_code_required.\n\n\"is_new_user\" is true only when THIS call created the account, so\nthe app can show first-run onboarding to genuinely new guests. It\nis returned by this endpoint only; no other token-pair response\ncarries it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify an OTP code",
                "parameters": [
                    {
                        "description": "Phone and code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.otpVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.otpVerifyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "invalid or expired code",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Exchanges a valid refresh token for a new access/refresh token pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.tokenPairResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "description": "Registers a new user with email + password and returns a token pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "New account details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.signupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.tokenPairResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "email already registered",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "s3cret123"
                }
            }
        },
        "auth.otpRequestRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "+77011234567"
                }
            }
        },
        "auth.otpRequestedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "populated only when AUTH_OTP_DEV_EXPOSE=true",
                    "type": "string",
                    "example": "123456"
                },
                "sent": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "auth.otpVerifyRequest": {
            "type": "object",
            "required": [
                "code",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "phone": {
                    "type": "string",
                    "example": "+77011234567"
                }
            }
        },
        "auth.otpVerifyResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6..."
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-07-12T18:30:00Z"
                },
                "is_new_user": {
                    "type": "boolean",
                    "example": true
                },
                "refresh_token": {
                    "type": "string",
                    "example": "9c8b7a6f-4d3e-2b1a-...-refresh"
                }
            }
        },
        "auth.refreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "9c8b7a6f-...-refresh"
                }
            }
        },
        "auth.signupRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "s3cret123"
                }
            }
        },
        "auth.tokenPairResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6..."
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-07-12T18:30:00Z"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "9c8b7a6f-4d3e-2b1a-...-refresh"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token from POST /api/v1/auth/login (or signup / otp/verify /\nrefresh), sent as: Authorization: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}